fix: info update location
