+++ src/main/java/com/spacestar/back/auth/service/MemberService.java
@@ -1,8 +1,12 @@
 package com.spacestar.back.auth.service;
 
+import com.spacestar.back.auth.dto.req.MemberInfoReqDto;
+
 public interface MemberService {
 
     void withdrawal(String uuid);
 
     void withdrawalForce(String uuid);
+
+    void updateMemberInfo(String uuid, MemberInfoReqDto memberInfoReqDto);
 }
